refactor(console): api import dialog - migrate material to MDC-based component
