+++ gravitee-apim-console-webui/src/management/api/component/gio-api-import-dialog/gio-api-import-dialog.component.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
 $typography: map.get(gio.$mat-legacy-theme, typography);
 
@@ -61,7 +61,6 @@ $typography: map.get(gio.$mat-legacy-theme, typography);
     padding-bottom: 8px;
 
     &__warn {
-      padding-left: 24px;
       @include mat.typography-level($typography, 'caption');
     }
