+++ gravitee-apim-console-webui/src/management/api/component/gio-api-import-dialog/gio-api-import-dialog.component.spec.ts
@@ -17,20 +17,20 @@ import { HarnessLoader, parallel } from '@angular/cdk/testing';
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { HttpTestingController } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { MatLegacyButtonHarness as MatButtonHarness } from '@angular/material/legacy-button/testing';
-import { MatLegacyCheckboxHarness as MatCheckboxHarness } from '@angular/material/legacy-checkbox/testing';
-import { MatLegacyDialogRef as MatDialogRef, MAT_LEGACY_DIALOG_DATA as MAT_DIALOG_DATA } from '@angular/material/legacy-dialog';
-import { MatLegacyFormFieldHarness as MatFormFieldHarness } from '@angular/material/legacy-form-field/testing';
+import { MatButtonHarness } from '@angular/material/button/testing';
+import { MatCheckboxHarness } from '@angular/material/checkbox/testing';
+import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatFormFieldHarness } from '@angular/material/form-field/testing';
 import { MatIconTestingModule } from '@angular/material/icon/testing';
-import { MatLegacyInputHarness as MatInputHarness } from '@angular/material/legacy-input/testing';
-import { MatLegacyTabHarness as MatTabHarness } from '@angular/material/legacy-tabs/testing';
+import { MatInputHarness } from '@angular/material/input/testing';
+import { MatTabHarness } from '@angular/material/tabs/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { GioFormFilePickerInputHarness } from '@gravitee/ui-particles-angular';
 
 import { GioApiImportDialogComponent } from './gio-api-import-dialog.component';
 import { GioApiImportDialogModule } from './gio-api-import-dialog.module';
 
-import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../testing';
+import { CONSTANTS_TESTING, GioHttpTestingModule } from '../../../../shared/testing';
 
 describe('GioApiImportDialogComponent', () => {
   let component: GioApiImportDialogComponent;
