+++ gravitee-apim-console-webui/src/management/api/general-info/api-general-info.component.ts
@@ -40,10 +40,7 @@ import { Constants } from '../../../entities/Constants';
 import { CategoryService } from '../../../services-ngx/category.service';
 import { PolicyService } from '../../../services-ngx/policy.service';
 import { SnackBarService } from '../../../services-ngx/snack-bar.service';
-import {
-  GioApiImportDialogComponent,
-  GioApiImportDialogData,
-} from '../../../shared/components/gio-api-import-dialog/gio-api-import-dialog.component';
+import { GioApiImportDialogComponent, GioApiImportDialogData } from '../component/gio-api-import-dialog/gio-api-import-dialog.component';
 import { GioPermissionService } from '../../../shared/components/gio-permission/gio-permission.service';
 import { ApiV2Service } from '../../../services-ngx/api-v2.service';
 import { Api, ApiV2, ApiV4, UpdateApi, UpdateApiV2, UpdateApiV4 } from '../../../entities/management-api-v2';
